sih-2024-main/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dlib @ file:///D:/sih-2024-main/sih-2024-main/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
face-recognition==1.3.0
face_recognition_models==0.3.0
idna==3.10
numpy==1.26.4
opencv-python==4.11.0.86
pillow==11.1.0
python-dotenv==1.1.0
requests==2.32.3
setuptools==78.0.2
urllib3==2.3.0
